backend/requirements.txt
fastapi==0.115.5
IMDbPY==2022.7.9
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
tmdbsimple==2.9.1
google-generativeai==0.3.2